% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment2param.R
\name{autoplot.moment2param}
\alias{autoplot.moment2param}
\title{autoplot.moment2param}
\usage{
\method{autoplot}{moment2param}(object, hjust = 0.5, ...)
}
\arguments{
\item{object}{..}

\item{hjust}{..}

\item{...}{..}
}
\description{
autoplot.moment2param
}
\keyword{internal}
