% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\docType{class}
\name{moment-class}
\alias{moment-class}
\title{Raw, Central and Standardized Moments}
\description{
..
}
\section{Slots}{

\describe{
\item{\code{distname}}{\link[base]{character} scalar, name of distribution,
e.g., \code{'norm'} for normal, \code{'sn'} for skew-normal, \code{'st'} for skew-\eqn{t},
and \code{'GH'} for Tukey \eqn{g}-&-\eqn{h} distribution,
following the nomenclature of \link[stats]{dnorm}, \link[sn]{dsn}, \link[sn]{dst} and \code{QuantileGH::dGH}}

\item{\code{location,scale}}{\link[base]{numeric} scalars or \link[base]{vector}s,
location and scale parameters}

\item{\code{raw1,raw2,raw3,raw4}}{\link[base]{numeric} scalars or \link[base]{vector}s,
1st or higher order \emph{raw} moments}

\item{\code{central2,central3,central4}}{\link[base]{numeric} scalars or \link[base]{vector}s,
2nd or higher order \emph{central} moments}

\item{\code{standardized3,standardized4}}{\link[base]{numeric} scalars or \link[base]{vector}s,
3rd or higher \emph{standardized} moments}
}}

\note{
Potential name clash with function \code{e1071::moment}.
}
