% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{moment2gamma}
\alias{moment2gamma}
\title{Solve Skew-Normal Parameters from Moments}
\usage{
moment2gamma(...)
}
\arguments{
\item{...}{\link[base]{numeric} scalars, \emph{\strong{two}} of
\code{mean} \eqn{\mu},
standard deviation \code{sd} \eqn{\sigma},
\code{skewness}, and
\code{kurtosis}}
}
\value{
Function \code{\link[=moment2gamma]{moment2gamma()}} returns a \link[base]{length}-2
\link[base]{numeric} \link[base]{vector} \eqn{(\alpha, \theta)}.
}
\description{
Solve skew-normal parameters from mean, standard deviation and skewness.
}
\details{
Function \code{\link[=moment2gamma]{moment2gamma()}} solves the
location \eqn{\xi}, scale \eqn{\omega} and slant \eqn{\alpha} parameters
of skew-normal distribution,
from user-specified mean \eqn{\mu} (default 0), standard deviation \eqn{\sigma} (default 1) and
skewness.
}
\keyword{internal}
