% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyGH.R
\name{moment2GH}
\alias{moment2GH}
\title{Solve Tukey \eqn{g}-&-\eqn{h} Parameters from Moments}
\usage{
moment2GH(mean = 0, sd = 1, skewness, kurtosis, g, h, ...)
}
\arguments{
\item{mean}{\link[base]{numeric} scalar, mean \eqn{\mu}, default value 0}

\item{sd}{\link[base]{numeric} scalar, standard deviation \eqn{\sigma}, default value 1}

\item{skewness}{\link[base]{numeric} scalar}

\item{kurtosis}{\link[base]{numeric} scalar}
}
\value{
Function \code{\link[=moment2GH]{moment2GH()}} returns a \link[base]{length}-4
\link[base]{numeric} \link[base]{vector} \eqn{(A, B, g, h)}.
}
\description{
Solve Tukey \eqn{g}-, \eqn{h}- and \eqn{g}-&-\eqn{h} distribution parameters
from mean, standard deviation, skewness and kurtosis.
}
\details{
Function \code{\link[=moment2GH]{moment2GH()}} solves the
location \eqn{A}, scale \eqn{B}, skewness \eqn{g}
and elongation \eqn{h} parameters of Tukey \eqn{g}-&-\eqn{h} distribution,
from user-specified mean \eqn{\mu} (default 0), standard deviation \eqn{\sigma} (default 1),
skewness and kurtosis.
}
\keyword{internal}
