% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{moment_gamma}
\alias{moment_gamma}
\title{Moments of Gamma Distribution}
\usage{
moment_gamma(scale = 1, shape)
}
\arguments{
\item{scale, shape}{\link[base]{numeric} scalars or \link[base]{vector}s,
scale \eqn{\theta} and shape \eqn{\alpha},
see function \link[stats]{dgamma}.}
}
\value{
Function \code{\link[=moment_gamma]{moment_gamma()}} returns a \linkS4class{moment} object.
}
\description{
Moments of Gamma Distribution
}
\keyword{internal}
