% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{moment_norm}
\alias{moment_norm}
\title{Moments of Normal Distribution}
\usage{
moment_norm(mean = 0, sd = 1)
}
\arguments{
\item{mean, sd}{\link[base]{numeric} scalars or \link[base]{vector}s,
see function \link[stats]{dnorm}}
}
\value{
Function \code{\link[=moment_norm]{moment_norm()}} returns a \linkS4class{moment} object.
}
\description{
Moments of \href{https://en.wikipedia.org/wiki/Normal_distribution}{normal distribution}.
}
\keyword{internal}
