% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn.R
\name{moment_sn}
\alias{moment_sn}
\title{Moments of Skew-Normal Distribution}
\usage{
moment_sn(xi = 0, omega = 1, alpha = 0)
}
\arguments{
\item{xi, omega, alpha}{\link[base]{numeric} scalars or \link[base]{vector}s,
location \eqn{\xi}, scale \eqn{\omega} and slant \eqn{\alpha},
see function \link[sn]{dsn}.}
}
\value{
Function \code{\link[=moment_sn]{moment_sn()}} returns a \linkS4class{moment} object.
}
\description{
Moments of Skew-Normal Distribution
}
\keyword{internal}
