% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{moment_st}
\alias{moment_st}
\title{Moments of Skew-\eqn{t} Distribution}
\usage{
moment_st(xi = 0, omega = 1, alpha = 0, nu = Inf)
}
\arguments{
\item{xi, omega, alpha, nu}{\link[base]{numeric} scalars or \link[base]{vector}s,
location \eqn{\xi}, scale \eqn{\omega}, slant \eqn{\alpha} and degree of freedom \eqn{\nu}
of function \link[sn]{dst}.}
}
\value{
Function \code{\link[=moment_st]{moment_st()}} returns a \linkS4class{moment} object.
}
\description{
Moments of Skew-\eqn{t} Distribution
}
\references{
Raw moments of skew-\eqn{t}: \url{https://arxiv.org/abs/0911.2342}
}
\keyword{internal}
