% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{show,moment-method}
\alias{show,moment-method}
\title{Show \linkS4class{moment}}
\usage{
\S4method{show}{moment}(object)
}
\arguments{
\item{object}{a \linkS4class{moment} object}
}
\value{
The \link[methods]{show} method for \linkS4class{moment} object
does not have a returned value.
}
\description{
Print S4 object \linkS4class{moment} in a pretty manner.
}
\keyword{internal}
