% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{moment_norm}
\alias{moment_norm}
\title{Moments of Normal Distribution}
\usage{
moment_norm(mean = 0, sd = 1)
}
\arguments{
\item{mean}{\link[base]{numeric} scalar or \link[base]{vector},
mean parameter \eqn{\mu}}

\item{sd}{\link[base]{numeric} scalar or \link[base]{vector},
standard deviation \eqn{\sigma}}
}
\value{
Function \code{\link[=moment_norm]{moment_norm()}} returns a \linkS4class{moment} object.
}
\description{
Moments of \href{https://en.wikipedia.org/wiki/Normal_distribution}{normal distribution}, parameter nomenclature follows
\link[stats]{dnorm} function.
}
\examples{
moment_norm(mean = 1.2, sd = .7)

}
